Fixed memory leak involving memory for audio paths When the application is terminated, all memory malloced in the creation of the app is freed. This includes the path of each of the paths for the audio.
When programming the freeing of the paths, the variable location of the path was written incorrectly so the memory was not freed.

This resulted in the first run of the application working fine, but when the app is terminated (not just paused), the memory wasn't freed. When the app is recreated (launced after termination), the program would crash because too much memory was taken up, resulting in a crash.

An issue that occured when debugging this resulted in trying to clean the memory mid-tests. Once I discovered the issue and before I realized it was a memory issue, I checked out the previous commits. Since the memory leak already occured on the main app, all runs even with the previous commits produced the same error, despite the old commits not creating additional memory issues. This made me believe briefly that this issue has existed for several commits, despite testing terminations in prior commits. The crash reports showed me that it was a memory leak rather than another issue. Based on the timing of the crash report, I assumed that the memory leak was on creation. Once the emulator was closed the previous memory issues cleared and it showed me that it occurs on the termination not the creation because the first creation of the app works fine.